requirements.txt
adal==1.2.0
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
azure==4.0.0
bcrypt==3.1.4
cached-property==1.5.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
cryptography==2.4.2
defusedxml==0.5.0
idna==2.7
isodate==0.6.0
lxml==4.3.0
more-itertools==4.3.0
msrest==0.6.2
msrestazure==0.5.1
oauthlib==2.1.0
paramiko==2.4.2
pathlib2==2.3.2
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.6.4
PyNaCl==1.3.0
python-dateutil==2.7.5
pywinrm==0.3.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
requests-credssp==1.1.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.24.1
zeep==3.2.0
